add pagination, further changes

+++ controllers/admin.js
@@ -1,6 +1,29 @@
 const { validationResult } = require("express-validator");
+const { deleteFile } = require("../utils/file");
 
 const Product = require("../models/product");
+
+const responseStatus422 = (
+	pageTitle,
+	editing,
+	product,
+	errorMessage,
+	validationResult,
+	req,
+	res,
+	next
+) => {
+	return res.status(422).render("admin/edit-product", {
+		pageTitle,
+		path: "/admin/edit-product",
+		editing,
+		hasError: true,
+		product,
+		isAuthenticated: req.user,
+		errorMessage,
+		validationResult
+	});
+};
 
 exports.getAdminAddProductPage = (req, res, next) => {
 	res.render("admin/edit-product", {
@@ -15,31 +38,37 @@ exports.getAdminAddProductPage = (req, res, next) => {
 };
 
 exports.postProduct = (req, res, next) => {
-	const { title, price, description, imageUrl } = req.body;
+	const { title, price, description } = req.body;
+	const image = req.file;
 	const errors = validationResult(req);
-
-	if (!errors.isEmpty()) {
-		return res.status(422).render("admin/edit-product", {
-			pageTitle: "Add product",
-			path: "/admin/edit-product",
-			editing: false,
-			hasError: true,
-			product: {
+	if (!errors.isEmpty() || !image) {
+		return responseStatus422(
+			"Add product",
+			false,
+			{
 				title,
 				price,
-				description,
-				imageUrl
+				description
 			},
-			isAuthenticated: req.user,
-			errorMessage: errors.array()[0].msg,
-			validationResult: errors
-		});
+			!image ? "Attached file is not an image" : errors.array()[0].msg,
+			!image ? [] : errors.array(),
+			req,
+			res,
+			next
+		);
 	}
-	const product = new Product({ title, price, description, imageUrl, userId: req.user._id });
+
+	const imageUrl = image.path;
+
+	const product = new Product({ title, price, description, image: imageUrl, userId: req.user._id });
 	product
 		.save()
 		.then(() => res.redirect("/admin/products"))
-		.catch(e => console.log(e));
+		.catch(e => {
+			const error = new Error(e);
+			error.httpStatusCode = 500;
+			return next(error);
+		});
 };
 
 exports.getProductToEdit = (req, res, next) => {
@@ -64,27 +93,27 @@ exports.getProductToEdit = (req, res, next) => {
 };
 
 exports.updateProduct = (req, res, next) => {
-	const { id, title, imageUrl, price, description } = req.body;
+	const { id, title, price, description } = req.body;
 
+	const image = req.file;
 	const errors = validationResult(req);
-	
+
 	if (!errors.isEmpty()) {
-		return res.status(422).render("admin/edit-product", {
-			pageTitle: "Edit product",
-			path: "/admin/edit-product",
-			editing: true,
-			hasError: true,
-			product: {
+		return responseStatus422(
+			"Add product",
+			true,
+			{
 				_id: id,
 				title,
 				price,
-				description,
-				imageUrl
+				description
 			},
-			isAuthenticated: req.user,
-			errorMessage: errors.array()[0].msg,
-			validationResult: errors
-		});
+			errors.array()[0].msg,
+			errors.array(),
+			req,
+			res,
+			next
+		);
 	}
 
 	Product.findById(id)
@@ -93,12 +122,19 @@ exports.updateProduct = (req, res, next) => {
 				return res.redirect("/");
 			}
 			product.title = title;
-			product.imageUrl = imageUrl;
+			if (image) {
+				deleteFile(product.image);
+				product.image = image.path;
+			}
 			product.price = price;
 			product.description = description;
 			return product.save().then(() => res.redirect("/admin/products"));
 		})
-		.catch(e => console.log(e));
+		.catch(e => {
+			const error = new Error(e);
+			error.httpStatusCode = 500;
+			return next(error);
+		});
 };
 
 exports.getAdminProducts = (req, res, next) => {
@@ -117,7 +153,19 @@ exports.getAdminProducts = (req, res, next) => {
 
 exports.deleteProduct = (req, res, next) => {
 	const { id } = req.params;
-	Product.deleteOne({ id, userId: req.user._id })
+	Product.findById(id)
+		.then(product => {
+			if (!product) {
+				return next(new Error("Product not found"));
+			}
+			deleteFile(product.image);
+			return Product.deleteOne({ id, userId: req.user._id });
+		})
 		.then(() => res.redirect("/admin/products"))
-		.catch(e => console.log(e));
+		.catch(e => {
+			const error = new Error(e);
+			error.httpStatusCode = 500;
+			return next(error);
+		})
+		.catch(e => next(e));
 };
